meaningful apis

+++ backend/ecom/src/main/java/com/cartsy/ecom/repository/ProductCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.cartsy.ecom.model.ProductCategory;
 
 
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long>{
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Integer>{
 
 }
